Move varying marker helper to utils and reuse it in slogdet
