+++ folx/ad.py
@@ -3,34 +3,12 @@ import jax.flatten_util as jfu
 import jax.numpy as jnp
 import jax.tree_util as jtu
 
+from .utils import _mark_varying_like
+
 
 def is_tree_complex(tree):
     leaves = jtu.tree_leaves(tree)
     return any(jnp.iscomplexobj(leaf) for leaf in leaves)
-
-
-def _varying_axes(x: jax.Array) -> tuple:
-    if not hasattr(jax, 'typeof'):
-        return ()
-
-    typ = jax.typeof(x)
-    manual_axis_type = getattr(typ, 'manual_axis_type', None)
-    if manual_axis_type is not None:
-        return tuple(getattr(manual_axis_type, 'varying', ()))
-
-    return tuple(getattr(typ, 'vma', ()))
-
-
-def _mark_varying_like(x: jax.Array, like: jax.Array) -> jax.Array:
-    axes = _varying_axes(like)
-    if not axes:
-        return x
-
-    if hasattr(jax.lax, 'pcast'):
-        return jax.lax.pcast(x, axes, to='varying')
-    if hasattr(jax.lax, 'pvary'):
-        return jax.lax.pvary(x, axes)
-    return x
 
 
 def vjp_rc(fun, *primals: jax.Array):
